Resolved compile-time warnings (VS 2013 only, still some warnings in gcc)

+++ src/spell.cpp
@@ -315,7 +315,7 @@ bool spellCast(unsigned int spell, int character, int param, SpellCastError *err
 		/* recalculate spell speed - based on 5/sec */
 		float MP_OF_LARGEST_SPELL = 45;
 		int spellMp = spells[spell].mp;
-		time = 10000.0 / settings.spellEffectSpeed  *  spellMp / MP_OF_LARGEST_SPELL;
+		time = int(10000.0 / settings.spellEffectSpeed  *  spellMp / MP_OF_LARGEST_SPELL);
 		soundPlay(SOUND_PREMAGIC_MANA_JUMBLE, false, time);
 		EventHandler::wait_msecs(time);
